+++ app/DI.Colef.Sia.Core/Resena.cs
@@ -147,7 +147,7 @@ namespace DecisionesInteligentes.Colef.Sia.Core
 
         public virtual int PaginaFinal { get; set; }
 
-        public virtual int Puntuacion { get; set; }
+        public virtual decimal Puntuacion { get; set; }
 
         [NotNull]
         public virtual Usuario Usuario { get; set; }
